+++ app/packages/klogs/src/components/panel/LogsPanel.tsx
@@ -1,107 +1,184 @@
-import { APIContext, APIError, timeOptions, ITimes, UseQueryWrapper, IPluginInstance, Pagination } from '@kobsio/core';
-import { Stack, Paper, Typography, Box } from '@mui/material';
-import { useQuery } from '@tanstack/react-query';
-import { FunctionComponent, useContext, useState } from 'react';
+import {
+  IPluginInstance,
+  PluginPanel,
+  PluginPanelActionLinks,
+  IPluginPanelProps,
+  PluginPanelError,
+} from '@kobsio/core';
+import { Paper, Tab, Tabs } from '@mui/material';
+import queryString from 'query-string';
+import { FunctionComponent, useState } from 'react';
 
-import { IQuery } from './Panel';
+import LogsQueryView from './LogsQueryView';
 
-import { ILogsData } from '../common/types';
-import LogsBucketChart, { IChangeTimeframePayload } from '../page/LogsBucketChart';
-import LogsTable from '../page/LogsTable';
+import isString from '../utils/isString';
 
-const now = () => Math.floor(Date.now() / 1000);
-
-const noopHandlers = {
-  // eslint-disable-next-line @typescript-eslint/no-empty-function
-  onAddFilter: () => {},
-  // eslint-disable-next-line @typescript-eslint/no-empty-function
-  onChangeSort: () => {},
-  // eslint-disable-next-line @typescript-eslint/no-empty-function
-  onSelectField: () => {},
-};
-
-interface ILogsQueryViewProps {
-  instance: IPluginInstance;
-  query: IQuery;
-  setTimes: (times: ITimes) => void;
-  times: ITimes;
+export interface IQuery {
+  fields: string[];
+  name: string;
+  query: string;
 }
 
-/**
- * LogsQueryView renders the results of a query in a table format
- * the component includes a bar chart for navigating the logs time-series
- */
-const LogsQueryView: FunctionComponent<ILogsQueryViewProps> = ({ instance, query, setTimes, times }) => {
-  const order = 'desc';
-  const orderBy = 'timestamp';
-  const { client } = useContext(APIContext);
-  const [page, setPage] = useState(1);
-  const [perPage, setPerPage] = useState(100);
+export interface IOptions {
+  queries: IQuery[];
+  type: 'logs';
+}
 
-  const queryResult = useQuery<ILogsData, APIError>([times], () => {
-    let timeEnd: number, timeStart;
-    if (times.time === 'custom') {
-      timeEnd = times.timeEnd;
-      timeStart = times.timeStart;
-    } else {
-      timeEnd = now();
-      timeStart = now() - timeOptions[times.time].seconds;
-    }
+interface ITabPanelProps {
+  children?: React.ReactNode;
+  index: number;
+  value: number;
+}
 
-    const path = `/api/plugins/klogs/logs?query=${query.query}&order=${order}&orderBy=${orderBy}&timeStart=${timeStart}&timeEnd=${timeEnd}`;
-    return client.get<ILogsData>(path, {
-      headers: {
-        'x-kobs-cluster': instance.cluster,
-        'x-kobs-plugin': instance.name,
-      },
-    });
-  });
-
-  const handleChangeTimeframe = (payload: IChangeTimeframePayload) => {
-    setTimes({
-      time: 'custom',
-      timeEnd: payload.timeEnd,
-      timeStart: payload.timeStart,
-    });
-  };
-
+const TabPanel: FunctionComponent<ITabPanelProps> = ({ index, value, children }) => {
   return (
-    <UseQueryWrapper
-      {...queryResult}
-      errorTitle="Failed to load applications"
-      isNoData={!queryResult.data || queryResult.data.documents === null}
-      noDataTitle="No logs found"
-      noDataMessage="There were no logs found for your search query"
-    >
-      <Stack direction="column" sx={{ width: '100%' }} spacing={2}>
-        <Stack direction="row" justifyContent="space-between" m={2}>
-          <Typography fontWeight="medium">{`${queryResult.data?.count} documents in ${queryResult.data?.took} milliseconds`}</Typography>
-        </Stack>
-        <Box sx={{ height: '150px' }}>
-          <LogsBucketChart buckets={queryResult.data?.buckets || []} onChangeTimeframe={handleChangeTimeframe} />
-        </Box>
-        <Paper>
-          <LogsTable
-            fields={query.fields}
-            options={{ hideActionColumn: true }}
-            order={order}
-            orderBy={orderBy}
-            rows={((rows) => rows.slice((page - 1) * perPage, page * perPage))(queryResult?.data?.documents || [])}
-            handlers={noopHandlers}
-          />
-        </Paper>
-        <Pagination
-          count={queryResult.data?.documents?.length ?? 0}
-          page={page ?? 1}
-          perPage={perPage}
-          handleChange={(page, perPage) => {
-            setPage(page);
-            setPerPage(perPage);
-          }}
-        />
-      </Stack>
-    </UseQueryWrapper>
+    <div role="tabpanel" hidden={value !== index} id={`tabpanel-${index}`} aria-labelledby={`tab-${index}`}>
+      {value === index && children}
+    </div>
   );
 };
 
-export default LogsQueryView;
+// utility for creating a uri that directs the user to the correct klogs instance, where
+// both the query and fields are already selected
+const uriFromQuery = (instance: IPluginInstance, query: IQuery) => {
+  const path = `/plugins/${instance.cluster}/klogs/${instance.name}`;
+  const search = queryString.stringify(
+    {
+      fields: query.fields,
+      query: query.query,
+    },
+    { arrayFormat: 'bracket', skipEmptyString: false, skipNull: false },
+  );
+
+  return [path, search].join('?');
+};
+
+// type guard for options
+// in future this could be checked with zod (https://zod.dev/)
+const isValid = (options?: IOptions): options is IOptions => {
+  if (!options) {
+    return false;
+  }
+
+  if (!Array.isArray(options.queries)) {
+    return false;
+  }
+
+  for (const query of options.queries) {
+    if (!Array.isArray(query.fields)) {
+      return false;
+    }
+    for (const field of query.fields) {
+      if (!isString(field)) {
+        return false;
+      }
+    }
+
+    if (!isString(query.query)) {
+      return false;
+    }
+
+    if (!isString(query.name)) {
+      return false;
+    }
+  }
+
+  return true;
+};
+
+/**
+ * Panel implements the panel for the klogs plugin
+ * it renders the queries in a table format
+ * depending on the queries, this component will either show a single table
+ * or display tabs, where one tab represents one query of the queries array
+ */
+const LogsPanel: FunctionComponent<IPluginPanelProps<IOptions>> = ({
+  description,
+  instance,
+  options,
+  setTimes,
+  times,
+  title,
+}) => {
+  const [tab, setTab] = useState(0);
+  if (!isValid(options)) {
+    return (
+      <PluginPanelError
+        title={title}
+        description={description}
+        message="Invalid options for KLogs plugin"
+        details="One of the required options: 'queries' is incorrect"
+        example={`plugin:
+  name: klogs
+  type: klogs
+  options:
+  queries:
+    - fields: ["app"]
+      query: "namespace='kube-system'`}
+        documentation="https://kobs.io/main/plugins/klogs/#panel-options"
+      />
+    );
+  }
+
+  // todo maybe this should also trogger the pluginpanelerror
+  if (options.queries.length === 0) {
+    return <></>;
+  }
+
+  if (options.queries.length === 1) {
+    const query = options.queries[0];
+    return (
+      <PluginPanel
+        title={title}
+        actions={
+          <PluginPanelActionLinks
+            links={[
+              {
+                link: uriFromQuery(instance, query),
+                title: `explore "${query.name}"`,
+              },
+            ]}
+            isFetching={false}
+          />
+        }
+      >
+        <LogsQueryView instance={instance} query={query} setTimes={setTimes} times={times} />
+      </PluginPanel>
+    );
+  }
+
+  return (
+    <PluginPanel
+      title={title}
+      actions={
+        <PluginPanelActionLinks
+          links={options.queries.map((query) => ({
+            link: uriFromQuery(instance, query),
+            title: `explore "${query.name}"`,
+          }))}
+          isFetching={false}
+        />
+      }
+    >
+      <Paper sx={{ height: '100%', width: '100%' }}>
+        <Tabs
+          value={tab}
+          onChange={(e, value: number) => setTab(value)}
+          aria-label="selects the search query"
+          sx={{ mb: 2 }}
+        >
+          {options.queries.map((query, i) => (
+            <Tab key={i} label={query.name} id={query.name} aria-controls={`tab-query-${i}`} />
+          ))}
+        </Tabs>
+        {options.queries.map((query, i) => (
+          <TabPanel key={i} index={i} value={tab}>
+            <LogsQueryView instance={instance} query={query} setTimes={setTimes} times={times} />
+          </TabPanel>
+        ))}
+      </Paper>
+    </PluginPanel>
+  );
+};
+
+export default LogsPanel;
