+++ app/packages/klogs/src/components/utils/isString.ts
@@ -1,0 +1,4 @@
+// shorthand for typeof comparison
+const isString = (v: unknown) => typeof v === 'string';
+
+export default isString;
